Refactor comment utilities into dedicated module
